Eliminated loop in audit_6.2.15.sh
The need is to list groups that are used but not defined. The old logic does a grep in /etc/group for every user. This is inefficent if there are many users. The new logic lists all the used groups, all the defined groups, and compares the lists.

I have seen /etc/passwd and /etc/group files with comments in them, but the old logic wouldn't handle them any better. One could pre-filter out comments.