ADDED: facts/all_mounts.fact
added a new fact that should be deployed to /etc/ansible/facts.d. The current ansible_mounts fact only will get mounts that are tied to a pyshical hardware device thereby ignoring tmpfs and potentialy other mounts that need option changes as part of the CIS hardening process. This fact will return all mounts in the facts and lets the task sort out which one it needs.